fix: address activity review feedback
