+++ apps/web/src/routes/_protected.workspaces/$workspaceId/-components/activity/activity-panel.logic.ts
@@ -1,121 +1,7 @@
 import { addDays, parseIsoDateLocal } from "@/lib/dates";
-import type {
-  MatterActivityFilters,
-  MatterActivityItem,
-} from "@/lib/workspaces/queries";
+import type { MatterActivityItem } from "@/lib/workspaces/queries";
 
 const DOCUMENT_BATCH_WINDOW_MS = 60_000;
-const FORMULA_PREFIX_RE = /^\s*[=+\-@\t\r\n]/u;
-
-const escapeCsv = (value: string): string => {
-  const formula = FORMULA_PREFIX_RE.test(value);
-  if (
-    !formula &&
-    !value.includes(",") &&
-    !value.includes('"') &&
-    !value.includes("\n") &&
-    !value.includes("\r")
-  ) {
-    return value;
-  }
-  const escaped = value.replace(/"/gu, '""');
-  return formula ? `"\t${escaped}"` : `"${escaped}"`;
-};
-
-export const matterActivityCsvRows = (
-  items: readonly MatterActivityItem[],
-): string[] => {
-  const rows = [
-    [
-      "Time",
-      "Actor",
-      "Actor type",
-      "Action",
-      "Category",
-      "Target type",
-      "Target",
-      "Origin",
-      "Run ID",
-      "Event ID",
-    ].join(","),
-  ];
-  for (const item of items) {
-    rows.push(
-      [
-        item.activityAt,
-        item.performer.name ?? "",
-        item.performer.type,
-        item.action,
-        item.category,
-        item.target.kind,
-        item.target.name ?? "",
-        [item.trigger.type, item.trigger.source].filter(Boolean).join(":"),
-        item.runId ?? "",
-        item.id,
-      ]
-        .map(escapeCsv)
-        .join(","),
-    );
-  }
-  return rows;
-};
-
-export const matterActivityJsonExport = ({
-  exportedAt,
-  filters,
-  items,
-}: {
-  exportedAt: string;
-  filters: MatterActivityFilters;
-  items: readonly MatterActivityItem[];
-}) => ({
-  version: 1 as const,
-  exportedAt,
-  filters,
-  items,
-});
-
-export const matterActivityExportResponse = ({
-  exportedAt,
-  filters,
-  format,
-  items,
-}: {
-  exportedAt: string;
-  filters: MatterActivityFilters;
-  format: "csv" | "json";
-  items: readonly MatterActivityItem[];
-}): Response => {
-  const chunks =
-    format === "csv"
-      ? matterActivityCsvRows(items).map((row) => `${row}\r\n`)
-      : [
-          JSON.stringify(
-            matterActivityJsonExport({ exportedAt, filters, items }),
-          ),
-        ];
-  const encoder = new TextEncoder();
-  let index = 0;
-  const body = new ReadableStream({
-    pull(controller) {
-      const chunk = chunks.at(index);
-      if (chunk === undefined) {
-        controller.close();
-        return;
-      }
-      index += 1;
-      controller.enqueue(encoder.encode(chunk));
-    },
-  });
-  return new Response(body, {
-    headers: {
-      "Content-Type":
-        format === "csv"
-          ? "text/csv; charset=utf-8"
-          : "application/json; charset=utf-8",
-    },
-  });
-};
 
 export const toMatterActivityDateRange = ({
   from,
